database

+++ PaymentGateway.Application/Queries/ListOfAccounts.cs
@@ -14,11 +14,11 @@ namespace PaymentGateway.Application.Queries
         public class Model
         {
             public int Id { get; set; }
-            public double Balance { get; set; }
+            public decimal Balance { get; set; }
             public string Currency { get; set; }
             public string Iban { get; set; }
             public string Status { get; set; }
-            public double Limit { get; set; }
+            public decimal Limit { get; set; }
             public string Type { get; set; }
         }
         public class Query : IRequest<List<Model>>
