+++ PaymentGateway.Models/Account.cs
@@ -2,19 +2,19 @@
 {
     public class Account
     {
-        public double Balance { get; set; }
+        public decimal Balance { get; set; }
         public string Currency { get; set; }
         public string IbanCode { get; set; }
         public string Type { get; set; }
         public string Status { get; set; }
-        public double Limit { get; set; }
+        public decimal Limit { get; set; }
         public int AccountID { get; set; }
         public int OwnerID { get; set; }
 
         public string OwnerCnp { get; set; }
     
 
-        public Account(double balance, string currency, string ibanCode, string type, string status, double limit, int id, string ownerCnp,int ownerID)
+        public Account(decimal balance, string currency, string ibanCode, string type, string status, decimal limit, int id, string ownerCnp,int ownerID)
         {
             Balance = balance;
             Currency = currency;
