+++ PaymentGateway.Models/Product.cs
@@ -4,9 +4,9 @@
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public double Value { get; set; }
+        public decimal Value { get; set; }
         public string Currency { get; set; }
-        public double Limit { get; set; }
+        public decimal Limit { get; set; }
 
     }
 }
