+++ PymentGateway.PublishedLanguage/Events/AccountCreated.cs
@@ -9,12 +9,12 @@ namespace PaymentGateway.PublishedLanguage.Events
         public string IbanCode { get; set; }
         public string Type { get; set; }
         public string Status { get; set; }
-        public double Limit { get; set; }
+        public decimal Limit { get; set; }
         public int AccountID { get; set; }
         public string OwnerCnp { get; set; }
 
 
-        public AccountCreated(double balance, string currency, string ibanCode, string type, string status, double limit, int accountID, string ownerCnp)
+        public AccountCreated(double balance, string currency, string ibanCode, string type, string status, decimal limit, int accountID, string ownerCnp)
         {
             this.Balance = balance;
             this.Currency = currency;
